refactor : JPA 관련 config 한 패키지로 묶음
